Move version file into modules
